Weather app structure and design

+++ app/src/main/java/com/danc/weatherdvt/data/services/OpenWeatherService.kt
@@ -8,9 +8,9 @@ import retrofit2.http.Query
 interface OpenWeatherService {
 
     @GET("data/2.5/weather")
-    suspend fun getCurrentLocationWeather(@Query("lat") latitude: Int, @Query("lon") longitude: Int) : CurrentLocationWeather
+    suspend fun getCurrentLocationWeather(@Query("lat") latitude: Double, @Query("lon") longitude: Double) : CurrentLocationWeather
 
     @GET("/data/2.5/forecast")
-    suspend fun getCurrentWeatherForecast(@Query("lat") latitude: Int, @Query("lon") longitude: Int): CurrentWeatherForecast
+    suspend fun getCurrentWeatherForecast(@Query("lat") latitude: Double, @Query("lon") longitude: Double): CurrentWeatherForecast
 
 }
